Extract ISO date validation to date-utils
Add a new packages/core/src/date-utils.ts exporting isValidIsoDate and remove duplicated implementations from document.ts and frontmatter.ts. Update both files to import the shared validator, reducing code duplication while preserving existing ISO date validation behavior.
