+++ packages/core/src/frontmatter.ts
@@ -1,5 +1,6 @@
 import { parse as parseYaml } from "yaml";
 import { DiagnosticError, lineRange } from "./diagnostics.js";
+import { isValidIsoDate } from "./date-utils.js";
 import type {
   FrontmatterDocument,
   PivotTableDefinition,
@@ -7,8 +8,6 @@ import type {
   SummaryRowDefinition,
   TableFrontmatter,
 } from "./types.js";
-
-const ISO_DATE_RE = /^\d{4}-\d{2}-\d{2}$/;
 
 function expectObject(value: unknown, context: string): Record<string, unknown> {
   if (typeof value === "object" && value !== null && !Array.isArray(value)) {
@@ -156,22 +155,6 @@ function parseReportTables(
   }
 
   return result;
-}
-
-function isValidIsoDate(value: string): boolean {
-  if (!ISO_DATE_RE.test(value)) return false;
-  const [yearRaw, monthRaw, dayRaw] = value.split("-");
-  const year = Number(yearRaw);
-  const month = Number(monthRaw);
-  const day = Number(dayRaw);
-  if (!Number.isInteger(year) || !Number.isInteger(month) || !Number.isInteger(day)) return false;
-
-  if (month < 1 || month > 12) return false;
-  if (day < 1) return false;
-
-  const isLeapYear = (year % 4 === 0 && year % 100 !== 0) || (year % 400 === 0);
-  const daysInMonth = [31, isLeapYear ? 29 : 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-  return day <= daysInMonth[month - 1];
 }
 
 function parsePivotTables(
